Add a uuid param for the Attraction.java class
Modify the type of the ticketType param from Map<String,Double> to Map<String,String> in the Attraction.java class

Add a AttractionDTO.java class to hold the data of an attraction.

Add a AttractionRepository.java interface to define the methods to interact with the Attraction data.

Add a AttractionService.java class to handle the business logic of the Attraction application.

Add a DataSetController.java to handle the HTTP requests and responses for the Information .

Add a DataSetService.java to handle the business logic of the Information and interact with Other Services.